requirements.txt
pandas~=1.0.5
matplotlib~=3.2.2
pytrends~=4.7.3
numpy~=1.19.0
phpserialize~=1.3
PyYAML~=5.3.1
PyMySQL~=0.9.3
requests~=2.24.0
pip~=20.1.1
lxml~=4.5.1
pytz~=2020.1
tqdm~=4.46.1
setuptools~=47.3.1
python-dateutil~=2.8.1
SQLAlchemy~=1.3.17
chardet~=3.0.4
six~=1.15.0
openpyxl~=3.0.3
jdcal~=1.4.1
idna~=2.9
urllib3~=1.25.9
certifi~=2019.11.28
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.2.0